+++ polars_baseball/__init__.py
@@ -29,7 +29,7 @@ from polars_baseball.enums.fangraphs import (
 )
 from polars_baseball.enums.position import Position
 
-__version__ = "0.20.1"
+__version__ = "0.20.2"
 
 
 _logging.getLogger("polars_baseball").addHandler(_logging.NullHandler())